% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/square.R
\name{square}
\alias{square}
\title{The function says it, dumbass}
\usage{
square(x)
}
\arguments{
\item{x}{vector you want to square}
}
\value{
numeric value - surprise, surprise - squared!
}
\description{
The function says it, dumbass
}
\examples{
square(5)
}
